refactor: moved ViewModels to a dedicated folder for better project structure
